Restructured api views folder and applied jobs api
