+++ jobsapp/api/views/__init__.py
@@ -1,0 +1,3 @@
+from .employee import *
+from .employer import *
+from .common import *
